refactor: convert Float usage to BigDecimal in whole codebase

+++ src/main/java/com/karifovas/gamerating/dto/AdjustmentDto.java
@@ -3,11 +3,13 @@ package com.karifovas.gamerating.dto;
 import com.karifovas.gamerating.model.AdjustmentType;
 import lombok.Builder;
 
+import java.math.BigDecimal;
+
 @Builder
 public record AdjustmentDto(
         String id,
         AdjustmentType type,
-        Float value,
+        BigDecimal value,
         Scale scale
 ) {
     public record Scale(
