+++ src/main/java/com/karifovas/gamerating/model/Factor.java
@@ -3,6 +3,8 @@ package com.karifovas.gamerating.model;
 import lombok.*;
 import lombok.experimental.FieldNameConstants;
 
+import java.math.BigDecimal;
+
 @Getter
 @Setter
 @Builder
@@ -13,7 +15,7 @@ public class Factor {
     private final String code;
     private final String name;
     private final String description;
-    private Float value;
+    private BigDecimal value;
 
     public Factor copy() {
         return Factor.builder()
