+++ src/main/java/com/karifovas/gamerating/repository/CustomFactorRepositoryImpl.java
@@ -10,6 +10,8 @@ import org.springframework.data.mongodb.core.query.Update;
 import org.springframework.stereotype.Repository;
 import reactor.core.publisher.Mono;
 
+import java.math.BigDecimal;
+
 @Repository
 @RequiredArgsConstructor
 public class CustomFactorRepositoryImpl implements FactorRepository, CustomFactorRepository {
@@ -41,7 +43,7 @@ public class CustomFactorRepositoryImpl implements FactorRepository, CustomFacto
     }
 
     @Override
-    public Mono<Boolean> updateValue(String id, String ratingId, String gameId, Float value) {
+    public Mono<Boolean> updateValue(String id, String ratingId, String gameId, BigDecimal value) {
         var query = new Query(
                 Criteria.where(Rating.Fields.id).is(ratingId)
                         .and(Rating.Fields.gameId).is(gameId)
